[Recovery Lifecycle][Auto-fix Worker Policy][D1] Split recovery worker from runtime (#1573)
## Summary

The recovery worker wrapper now lives outside the generic runtime
module.

The moved wrapper remains behavior-neutral and defaults to a no-op tick.

## Test Plan

- [x] `pnpm --filter @invoker/app exec vitest run
src/__tests__/worker-runtime.test.ts
src/__tests__/auto-fix-recovery.test.ts`

## Revert Plan

- Safe to revert? Yes
- Revert command: `git revert <merge-commit-sha>`
- Post-revert steps: None
- Data migration? No
