+++ packages/app/src/__tests__/worker-runtime.test.ts
@@ -1,9 +1,7 @@
 import { afterEach, describe, expect, it, vi } from 'vitest';
 
 import {
-  createRecoveryWorker,
   createWorkerRuntime,
-  RECOVERY_WORKER_KIND,
   type WorkerTickContext,
 } from '../worker-runtime.js';
 
@@ -361,39 +359,6 @@ describe('worker runtime', () => {
       expect(process.listenerCount('SIGINT')).toBe(before + 1);
       await runtime.stop();
       expect(process.listenerCount('SIGINT')).toBe(before);
-    });
-  });
-
-  describe('recovery worker', () => {
-    it('exposes the recovery identity', () => {
-      const runtime = createRecoveryWorker({ logger, instanceId: 'rec-1', installSignalHandlers: false });
-      expect(runtime.identity).toEqual({ kind: RECOVERY_WORKER_KIND, instanceId: 'rec-1' });
-    });
-
-    it('is behavior-neutral: its default tick does nothing and does not throw', async () => {
-      const runtime = createRecoveryWorker({ logger, instanceId: 'rec-2', installSignalHandlers: false });
-      await expect(runtime.tick()).resolves.toBeUndefined();
-      expect(logger.error).not.toHaveBeenCalled();
-      await runtime.stop();
-    });
-
-    it('does not auto-run a tick on start', async () => {
-      vi.useFakeTimers();
-      const onTick = vi.fn();
-      const runtime = createRecoveryWorker({
-        logger,
-        instanceId: 'rec-3',
-        intervalMs: 1000,
-        onTick,
-        installSignalHandlers: false,
-      });
-      runtime.start();
-      await Promise.resolve();
-      // tickOnStart defaults to false for the recovery worker.
-      expect(onTick).not.toHaveBeenCalled();
-      await vi.advanceTimersByTimeAsync(1000);
-      expect(onTick).toHaveBeenCalledTimes(1);
-      await runtime.stop();
     });
   });
 });
